+++ zarrs/src/array/codec/array_to_bytes/packbits/packbits_codec.rs
@@ -1,5 +1,8 @@
 #![allow(clippy::similar_names)]
 
+// FIXME: This codec was really hacked together.
+// It can probably be written much cleaner and with simpler logic.
+
 use std::{borrow::Cow, sync::Arc};
 
 use num::Integer;
@@ -319,15 +322,20 @@ impl ArrayToBytesCodecTraits for PackBitsCodec {
             }
             if sign_extension {
                 let signed: bool = {
+                    let bit_enc0 = component_idx * component_size_bits_extracted;
+                    let (byte_enc, bit_enc) =
+                        (bit_enc0 + component_size_bits_extracted.saturating_sub(1)).div_rem(&8);
+                    ((packed_elements[usize::try_from(byte_enc).unwrap()] >> bit_enc) & 0b1) == 1
+                };
+                if signed {
                     let (byte_dec, bit_dec) = div_rem_8bit(
                         bit_dec0 + component_size_bits_extracted.saturating_sub(1),
                         component_size_bits,
                     );
-                    bytes_dec[usize::try_from(byte_dec).unwrap()] >> bit_dec & 0x1 == 1
-                };
-                if signed {
-                    for bit in component_size_bits_extracted..component_size_bits {
-                        let (byte_dec, bit_dec) = div_rem_8bit(bit_dec0 + bit, component_size_bits);
+                    // Sign-extend to all remaining bits in the byte
+                    // This differs from the spec which says sign extend to N (component_size_bits) bits
+                    // This makes it just work with int4 / int2 -> int8
+                    for bit_dec in bit_dec + 1..8 {
                         bytes_dec[usize::try_from(byte_dec).unwrap()] |= 1 << bit_dec;
                     }
                 }